PL-315 Added new datatypes file
